3.8.68.550 patched a bug where pre-recorded message can get deleted after reading improved stability

+++ Re_TTSCat/TTSHelper/TTSPlayer.Play.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using System.Threading;
 using System.Windows.Threading;
 using NAudio.Wave;
@@ -8,37 +9,47 @@ namespace Re_TTSCat
 {
     public static partial class TTSPlayer
     {
-        public static void Play(string filename, bool wait = true)
+        public static void Play(string filename, bool wait = true, bool forceKeepCache = false)
         {
             var frame = new DispatcherFrame();
             var thread = new Thread(() =>
             {
-                using (var reader = new AudioFileReader(filename))
+                try
                 {
-                    using (var waveOut = new WaveOutEvent())
+                    using (var reader = new AudioFileReader(filename))
                     {
-                        waveOut.Init(reader);
-                        reader.Volume = Volume;
-                        Bridge.ALog($"音量设置为: {Volume}");
-                        waveOut.Play();
-                        Vars.TotalPlayed++;
-                        if (!wait)
-                            frame.Continue = false;
-                        while (waveOut.PlaybackState != PlaybackState.Stopped)
+                        using (var waveOut = new WaveOutEvent())
                         {
-                            if (Vars.CallPlayerStop)
-                                waveOut.Stop();
-                            if (!reader.Volume.IsNearEnough(Volume, 0.02f))
+                            waveOut.Init(reader);
+                            reader.Volume = Volume;
+                            Bridge.ALog($"音量设置为: {Volume}");
+                            waveOut.Play();
+                            Vars.TotalPlayed++;
+                            if (!wait)
+                                frame.Continue = false;
+                            while (waveOut.PlaybackState != PlaybackState.Stopped)
                             {
-                                Bridge.ALog($"同步音量: {Volume}");
-                                reader.Volume = Volume;
+                                if (Vars.CallPlayerStop)
+                                    waveOut.Stop();
+                                if (!reader.Volume.IsNearEnough(Volume, 0.02f))
+                                {
+                                    Bridge.ALog($"同步音量: {Volume}");
+                                    reader.Volume = Volume;
+                                }
+                                Thread.Sleep(50);
                             }
-                            Thread.Sleep(50);
                         }
                     }
+                    if (Vars.CurrentConf.DoNotKeepCache && !forceKeepCache)
+                    {
+                        Bridge.ALog($"正在删除缓存文件: {filename}");
+                        File.Delete(filename);
+                    }
                 }
-                if (Vars.CurrentConf.DoNotKeepCache)
-                    File.Delete(filename);
+                catch (Exception ex)
+                {
+                    Bridge.ALog($"播放过程中发生错误: {Path.GetFileName(filename)}: {ex.Message}");
+                }
                 frame.Continue = false;
             });
             thread.SetApartmentState(ApartmentState.STA);
